+++ relay-config/src/config.rs
@@ -1137,9 +1137,6 @@ pub struct Processing {
     pub max_session_secs_in_past: u32,
     /// Kafka producer configurations.
     pub kafka_config: Vec<KafkaConfigParam>,
-    /// Configure what span format to produce.
-    #[serde(default)]
-    pub span_producers: SpanProducers,
     /// Additional kafka producer configurations.
     ///
     /// The `kafka_config` is the default producer configuration used for all topics. A secondary
@@ -1197,36 +1194,6 @@ impl Default for Processing {
             attachment_chunk_size: default_chunk_size(),
             projectconfig_cache_prefix: default_projectconfig_cache_prefix(),
             max_rate_limit: default_max_rate_limit(),
-            span_producers: Default::default(),
-        }
-    }
-}
-
-/// Configuration for span producers.
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(default)]
-pub struct SpanProducers {
-    /// Random sample of organizations to produce json and no protobuf.
-    ///
-    /// Overrides both `produce_json` and `produce_protobuf` for matching orgs.
-    pub produce_json_sample_rate: Option<f32>,
-    /// List of organization IDs to produce JSON spans for.
-    ///
-    /// Overrides both `produce_json` and `produce_protobuf` for matching orgs.
-    pub produce_json_orgs: Vec<u64>,
-    /// Send JSON spans to `ingest-spans`.
-    pub produce_json: bool,
-    /// Send Protobuf (TraceItem) to `snuba-items`.
-    pub produce_protobuf: bool,
-}
-
-impl Default for SpanProducers {
-    fn default() -> Self {
-        Self {
-            produce_json_sample_rate: None,
-            produce_json_orgs: vec![],
-            produce_json: false,
-            produce_protobuf: true,
         }
     }
 }
@@ -2631,11 +2598,6 @@ impl Config {
     pub fn accept_unknown_items(&self) -> bool {
         let forward = self.values.routing.accept_unknown_items;
         forward.unwrap_or_else(|| !self.processing_enabled())
-    }
-
-    /// Returns the configuration for span producers.
-    pub fn span_producers(&self) -> &SpanProducers {
-        &self.values.processing.span_producers
     }
 }